nada

+++ src/main/java/com/utn/tacs/tp2016c1g4/marvel_webapp/api/recursos/PerfilResource.java
@@ -29,7 +29,7 @@ public class PerfilResource {
 	public Response get(@PathParam("username") String userName) {
 		logger.debug("get invocado");
 		PerfilGetResponse response = new PerfilGetResponse();
-		response.setId(1);
+		response.setId((long) 1);
 		response.setUsername(userName);
 		return Response.status(200).entity(response).build();
 	}
